+++ app/controllers/Substances.java
@@ -6,6 +6,7 @@ import models.User;
 import play.data.Form;
 import play.mvc.Controller;
 import play.mvc.Result;
+import play.mvc.Results;
 import play.mvc.Security;
 
 /**
@@ -15,6 +16,10 @@ import play.mvc.Security;
 public class Substances extends Controller {
 
     public static Result add() {
+        User user = User.find.byId(request().username());
+        if (!user.isAdmin) {
+            return Results.forbidden("Need to be admin");
+        }
         final Form<Substance> substanceForm = Form.form(Substance.class).bindFromRequest();
         final Substance substance = substanceForm.get();
 
@@ -23,16 +28,22 @@ public class Substances extends Controller {
     }
 
     public static Result edit(Long id) {
+        User user = User.find.byId(request().username());
+        if (!user.isAdmin) {
+            return Results.forbidden("Need to be admin");
+        }
         Substance substance = Substance.find.where().idEq(id).findUnique();
         Form<Substance> editForm = Form.form(Substance.class).fill(substance);
-        User user = User.find.byId(request().username());
         return ok(views.html.Substance.edit.render(substance, editForm, user));
     }
 
     public static Result update(Long id) {
+        User user = User.find.byId(request().username());
+        if (!user.isAdmin) {
+            return Results.forbidden("Need to be admin");
+        }
         Form<Substance> filledForm = Form.form(Substance.class).bindFromRequest();
         if (filledForm.hasErrors()) {
-            User user = User.find.byId(request().username());
             Substance substance = Substance.find.where().idEq(id).findUnique();
             return badRequest(views.html.Substance.edit.render(substance, filledForm, user));
         } else {
@@ -44,6 +55,10 @@ public class Substances extends Controller {
     }
 
     public static Result delete(Long id) {
+        User user = User.find.byId(request().username());
+        if (!user.isAdmin) {
+            return Results.forbidden("Need to be admin");
+        }
         final Substance substance = Substance.find.byId(id);
         if (substance != null) {
             substance.delete();
